Group grudger strategies together
